% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_depth.R
\name{extract_depth}
\alias{extract_depth}
\title{Extract Depth from a File Name}
\usage{
extract_depth(file.name)
}
\arguments{
\item{file.name}{Character string specifying the file name or path from which to
extract the depth (e.g., "T2_data_2m.csv").}
}
\value{
A character string containing the first matched depth (e.g., "2m") or
\code{NA} if no depth pattern is found in the file name.
}
\description{
Extracts a depth substring (e.g., 1m, 2m, 6m, 8m, 14m, 20m, 23m, 30m, 40m, 54m)
from a given file name. The function searches for the first occurrence of a
predefined depth pattern in the file name and returns it. If no depth is found,
it returns NA.
}
\examples{
# Example usage
extract_depth("T2_temperature_2m_2023.csv")  # Returns "2m"
extract_depth("t2_no_depth.csv")             # Returns NA

}
